requirements.txt
emoji>=1.2.0certifi==2020.12.5
chardet==4.0.0
click==7.1.2
docopt==0.6.2
emoji==1.2.0
filelock==3.0.12
idna==2.10
joblib==1.0.1
numpy==1.20.1
oauthlib==3.1.0
packaging==20.9
Pillow==8.1.2
pyparsing==2.4.7
PySocks==1.7.1
python-dotenv==0.15.0
regex==2020.11.13
requests==2.25.1
requests-oauthlib==1.3.0
sacremoses==0.0.43
six==1.15.0
tokenizers==0.10.1
torch==1.8.0+cu111
torchaudio==0.8.0
torchvision==0.9.0+cu111
tqdm==4.59.0
transformers==4.3.3
tweepy==3.10.0
typing-extensions==3.7.4.3
urllib3==1.26.3
